Auction _ flask/requirements.txt
bcrypt==4.1.3
blinker==1.8.2
cachelib==0.13.0
click==8.1.7
flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
flask-session==0.8.0
flask-sqlalchemy==3.1.1
greenlet==3.0.3
importlib-metadata==8.0.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
msgspec==0.18.6
PyJWT==2.8.0
SQLAlchemy==2.0.31
typing-extensions==4.12.2
werkzeug==3.0.3
zipp==3.19.2
